♻️ refactor: entities/organization 신설, challenger 도메인 타입 분해 및 re-export shim 제거
[변경 내용]
- entities/organization 신설: api/organization + model/types(기수/지부/학교 8개 타입)
- shared/model/domain 신설: 도메인 공통 enum 6개(Part, RoleType, ChallengerStatus,
  MemberStatus, OrganizationType, PointType)를 shared kernel로 승격
- entities/member 보강: api/member(회원 검색) 이동, model/people(Search*/Member*
  12개 타입) 신설. challenger.ts는 응답 타입 3개만 남기고 enum은 shared에서 re-export
- features/challenger/model/types는 grant/record 액션 타입 3개만 남김,
  이전 커밋의 하위 호환 re-export shim 완전 제거
- challenger/model/types 외부 importer 17개를 심볼별로 올바른 출처로 재지정

[변경 이유]
- 학교/지부/기수는 matching, application, project 등 여러 화면이 공유하는 조직
  도메인 데이터이므로 entities/organization이 올바른 위치
- enum을 shared로 올린 핵심 이유: member와 organization은 형제 entity라 서로
  import하면 경계 위반. 공통 어휘를 최하층 shared kernel에 두어야 양쪽이 위반 없이
  참조 가능. shared/view-mode의 Part 참조도 shared 내부 참조가 되어 위반 해소
- boundaries 린트 위반 122 -> 107건 (-15)

[고려 사항]
- api/member를 organization이 아닌 member로 배치(계획 9-2 대비 조정): searchMembers
  등은 조직 구조가 아니라 "사람"을 조회하므로 member 도메인이 정확
- MemberInfoResponse가 2개(entities/member/api/me의 현재 사용자용 vs
  model/people의 admin 조회용) 공존. 서로 다른 파일/스키마라 배럴 없이 분리 유지
- grant/record 타입(포인트 부여, 기록 코드)은 admin feature 액션이므로
  features/challenger에 잔류
- shared -> entities 17건(view-mode의 MemberInfoResponse/ChallengerInfoResponse
  참조)은 shared 역방향 정리 단계에서 해소 예정
- vi.mock의 문자열 경로는 import 자동 치환에 안 걸려 별도 수정(matching 테스트 2건)
- ProjectApplyModal.ui.test 5건 실패는 리팩토링 이전부터의 기존 이슈로 무관
