+++ src/entities/member/model/challenger.ts
@@ -1,59 +1,22 @@
-// 회원(챌린저) 도메인 공통 타입. 기수/파트/역할/상태 등
-// 여러 슬라이스가 공유하는 어휘라 entities/member에 둔다.
+// 챌린저(회원) 도메인 응답 타입. 공통 enum은 shared/model/domain에서 가져와 re-export.
 
-export type Part =
-  | "PLAN"
-  | "DESIGN"
-  | "WEB"
-  | "ANDROID"
-  | "IOS"
-  | "NODEJS"
-  | "SPRINGBOOT"
-  | "ADMIN"
+export type {
+  ChallengerStatus,
+  MemberStatus,
+  OrganizationType,
+  Part,
+  PointType,
+  RoleType,
+} from "@/shared/model/domain"
 
-export type ChallengerStatus = "ACTIVE" | "GRADUATED" | "EXPELLED" | "WITHDRAWN"
-
-export type MemberStatus = "ACTIVE" | "INACTIVE" | "WITHDRAWN"
-
-export type OrganizationType = "CENTRAL" | "CHAPTER" | "SCHOOL"
-
-export type RoleType =
-  | "CHALLENGER"
-  | "SUPER_ADMIN"
-  | "CENTRAL_PRESIDENT"
-  | "CENTRAL_VICE_PRESIDENT"
-  | "CENTRAL_OPERATING_TEAM_MEMBER"
-  | "CENTRAL_EDUCATION_TEAM_MEMBER"
-  | "CHAPTER_PRESIDENT"
-  | "SCHOOL_PRESIDENT"
-  | "SCHOOL_VICE_PRESIDENT"
-  | "SCHOOL_PART_LEADER"
-  | "SCHOOL_ETC_ADMIN"
-
-export type PointType =
-  | "BEST_WORKBOOK"
-  | "WARNING"
-  | "OUT"
-  | "CUSTOM"
-  | "BLOG_CHALLENGE"
-  | "BEST_WORKBOOK_V2"
-  | "UMC_EVENT_REVIEW"
-  | "PEER_REVIEW_SUBMISSION"
-  | "NO_WORKBOOK_MISSION"
-  | "STUDY_LATE"
-  | "STUDY_ABSENT"
-  | "EVENT_LATE"
-  | "EVENT_EARLY_LEAVE"
-  | "EVENT_LATE_CANCEL"
-  | "EVENT_NO_SHOW"
-  | "PART_LEAD_FEEDBACK_LATE"
-  | "SCHOOL_CORE_MEETING_ABSENT"
-  | "SCHOOL_CORE_TASK_NOT_COMPLETED"
-
-/**
- * 백엔드는 모든 ID 류 (`*Id`) 를 Java Long 정밀도 보전을 위해 numeric string 으로 직렬화한다.
- * FE 도 ID 는 string 으로 다룬다.
- */
+import type {
+  ChallengerStatus,
+  MemberStatus,
+  OrganizationType,
+  Part,
+  PointType,
+  RoleType,
+} from "@/shared/model/domain"
 
 export interface ChallengerPointInfo {
   id: string
