+++ src/features/challenger/api/challengerPoint.ts
@@ -1,9 +1,7 @@
 import { api } from "@/shared/lib/axios"
 
-import type {
-  ChallengerInfoResponse,
-  GrantChallengerPointRequest,
-} from "@/features/challenger/model/types"
+import type { ChallengerInfoResponse } from "@/entities/member/model/challenger"
+import type { GrantChallengerPointRequest } from "@/features/challenger/model/types"
 import type { ApiResponse } from "@/shared/lib/apiResponse"
 
 export async function grantChallengerPoints(
