+++ src/features/challenger/index.ts
@@ -1,3 +1,2 @@
-export { type RoleType } from "./model/types"
 export { GrantPointsPage } from "./ui/grant-points/GrantPointsPage"
 export { RecordCodePage } from "./ui/record-code/RecordCodePage"
