+++ src/features/challenger/ui/grant-points/MemberSearchPanel.tsx
@@ -1,13 +1,13 @@
 import { keepPreviousData, useQuery } from "@tanstack/react-query"
 import { useEffect, useRef, useState } from "react"
 
-import { searchMembers } from "@/features/challenger/api/member"
+import { searchMembers } from "@/entities/member/api/member"
 import { PART_LABEL } from "@/features/challenger/model/enums"
 import SearchIcon from "@/shared/assets/icon/search/SearchIcon"
 import { cn } from "@/shared/lib/utils"
 import { InputBox } from "@/shared/ui/input/InputBox"
 
-import type { SearchMemberItem } from "@/features/challenger/model/types"
+import type { SearchMemberItem } from "@/entities/member/model/people"
 
 interface MemberSearchPanelProps {
   selectedMember: SearchMemberItem | null
