+++ src/features/notice/model/defaultNotices.test.ts
@@ -6,9 +6,8 @@ import type { MemberInfoResponse } from "@/entities/member/api/me"
 import type {
   ChallengerInfoResponse,
   ChallengerRoleResponse,
-  Part,
-  RoleType,
-} from "@/features/challenger/model/types"
+} from "@/entities/member/model/challenger"
+import type { Part, RoleType } from "@/shared/model/domain"
 
 function makeMe({
   part,
