+++ src/features/project/list/model/matchingProjectList.ts
@@ -3,7 +3,7 @@ import { useNavigate, useSearch } from "@tanstack/react-router"
 import { useCallback, useEffect, useMemo, useState } from "react"
 
 import { isOperator } from "@/entities/member/model/identity"
-import { getChaptersWithSchools } from "@/features/challenger/api/organization"
+import { getChaptersWithSchools } from "@/entities/organization/api/organization"
 import { projectKeys } from "@/features/project/new/api"
 import { useActiveGisu } from "@/shared/hooks/useActiveGisu"
 import { formatSchoolName } from "@/shared/lib/formatSchoolName"
