+++ src/routes/matching/projects/announce/index.tsx
@@ -12,11 +12,11 @@ import {
   isChapterPresident,
   isSuperAdmin,
 } from "@/entities/member/model/identity"
-import { ensureMe } from "@/features/auth/lib/ensureMe"
 import {
   getAllChapters,
   getAllGisu,
-} from "@/features/challenger/api/organization"
+} from "@/entities/organization/api/organization"
+import { ensureMe } from "@/features/auth/lib/ensureMe"
 import {
   type Chapter,
   CHAPTERS,
