+++ src/shared/view-mode/projectViewContext.ts
@@ -4,7 +4,7 @@ import { getCurrentGisuChallengerRecords } from "./currentGisuRecords"
 import { isOtherChallengerPart, isPmPart } from "./viewModeParts"
 
 import type { MemberInfoResponse } from "@/entities/member/api/me"
-import type { Part } from "@/features/challenger/model/types"
+import type { Part } from "@/shared/model/domain"
 
 import type { ViewMode } from "."
